Add support for TOML formatted /recipes/new and /recipes/info
The native format of a recipe is TOML, this adds the ability to add a
new recipe, and to retrieve individual recipes as TOML results instead
of JSON.

To request a recipe as toml:
    curl http://HOST/api/v0/recipes/info/<name>?format=toml

To add a new TOML formatted recipe:
    curl -H "Content-Type: text/x-toml" -X POST --data-binary @nginx.toml http://HOST/api/v0/recipes/new

This uses Rocket's FromData trait to deserialize a text/x-toml POST into
the Recipe struct. It does require that the data be passed in binary
mode so that CR/LF isn't disturbed.